fix missing file issues and weird script calling, add topdeck.gg support
