+++ .gitignore
@@ -1,3 +1,5 @@
+temp/*
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
